+++ packages/rschedule/src/lib/operators/add.operator.ts
@@ -31,7 +31,7 @@ export class AddOperator<T extends typeof DateAdapter> extends Operator<T> {
   }
 
   /** Not actually used but necessary for IRunnable interface */
-  set(_: 'timezone', value: string | undefined) {
+  set(_: 'timezone', value: string | null) {
     return new AddOperator(this._streams.map(stream => stream.set('timezone', value)), {
       ...this.config,
       base: this.config.base && this.config.base.set('timezone', value),