+++ packages/rschedule/src/lib/operators/unique.operator.ts
@@ -23,7 +23,7 @@ export class UniqueOperator<T extends typeof DateAdapter> extends Operator<T> {
   }
 
   /** Not actually used but necessary for IRunnable interface */
-  set(_: 'timezone', value: string | undefined) {
+  set(_: 'timezone', value: string | null) {
     return new UniqueOperator([], {
       ...this.config,
       base: this.config.base && this.config.base.set('timezone', value),