+++ packages/rschedule/src/lib/rschedule-config.ts
@@ -2,6 +2,6 @@ import { DateAdapter } from './date-adapter';
 
 export class RScheduleConfig {
   static defaultDateAdapter: typeof DateAdapter | undefined;
-  static defaultTimezone: string | undefined;
+  static defaultTimezone: string | null = null;
   static defaultMaxFailedIterations: number | undefined;
 }